Move the interface into more generic part. Also, move the `EditModeSection` with edit move injection with more generic part because technically it need to change the edit mode outside by injection.
