+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Settings/LyricEditorEditModeSection.cs
@@ -8,39 +8,22 @@ using osu.Framework.Allocation;
 using osu.Framework.Localisation;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Settings.Components.Markdown;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.States;
-using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.States.Modes;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Components.Markdown;
 
 namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Settings
 {
-    public abstract partial class LyricEditorEditModeSection<TEditModeState, TEditMode> : LyricEditorEditModeSection<TEditMode>
+    public abstract partial class LyricEditorEditModeSection<TEditModeState, TEditMode> : EditModeSection<TEditModeState, TEditMode>
         where TEditModeState : IHasEditModeState<TEditMode> where TEditMode : Enum
-    {
-        [Resolved]
-        private TEditModeState tEditModeState { get; set; }
-
-        protected sealed override TEditMode DefaultMode() => tEditModeState.EditMode;
-
-        internal sealed override void UpdateEditMode(TEditMode mode)
-        {
-            tEditModeState.ChangeEditMode(mode);
-
-            base.UpdateEditMode(mode);
-        }
-    }
-
-    public abstract partial class LyricEditorEditModeSection<TEditMode> : EditModeSection<TEditMode>
-        where TEditMode : Enum
     {
         protected sealed override LocalisableString Title => "Edit mode";
 
         [Resolved]
         private ILyricSelectionState lyricSelectionState { get; set; }
 
-        protected override DescriptionTextFlowContainer CreateDescriptionTextFlowContainer()
+        protected sealed override DescriptionTextFlowContainer CreateDescriptionTextFlowContainer()
             => new LyricEditorDescriptionTextFlowContainer();
 
-        internal override void UpdateEditMode(TEditMode mode)
+        internal sealed override void UpdateEditMode(TEditMode mode)
         {
             // should cancel the selection after change to the new edit mode.
             lyricSelectionState?.EndSelecting(LyricEditorSelectingAction.Cancel);
